+++ author.php
@@ -6,7 +6,7 @@
 
 <?php get_header(); ?>
 
-	<section class="archive author-archive">
+	<main class="archive author-archive">
 		<div class="a_content">
 			<div class="a_content_inner">
 				<?php if ( have_posts() ) : ?>
@@ -52,7 +52,7 @@
 				<?php endif; ?>
 			</div>
 		</div>
-	</section>
+	</main>
 
 <?php get_sidebar(); ?>
 <?php get_footer(); ?>