+++ page.php
@@ -6,12 +6,12 @@
 
 <?php get_header(); ?>
 
-<div class="a_content">
+<main class="a_content">
 	<?php the_post();?>
 
 	<?php get_template_part( 'content', 'page' ); ?>
 
 	<?php //comments_template( '', true ); ?>
-</div>
+</main>
 
 <?php get_footer(); ?>