+++ sidebar.php
@@ -6,7 +6,7 @@
 
 <?php if ( !is_internal() && 'post' == get_post_type() ) : ?>
 	<?php if ( is_active_sidebar( 'sidebar-blog' ) ) : ?>
-		<div class="a_sidebar">
+		<aside class="a_sidebar">
 			<div class="a_sidebar_inner">
 				<hr />
 				<aside class="sidebar left">
@@ -14,6 +14,6 @@
 				</aside>
 				<div class="clear"></div>
 			</div>
-		</div>
+		</aside>
 	<?php endif; ?>
 <?php endif; ?>